media state: extract media thunks into their own file
