+++ client/post-editor/media-modal/gallery/caption.jsx
@@ -11,7 +11,7 @@ import { connect } from 'react-redux';
  * Internal dependencies
  */
 import FormTextInput from 'components/forms/form-text-input';
-import { updateMedia } from 'state/media/actions';
+import { updateMedia } from 'state/media/thunks';
 
 class EditorMediaModalGalleryCaption extends React.Component {
 	static displayName = 'EditorMediaModalGalleryCaption';
@@ -53,7 +53,7 @@ class EditorMediaModalGalleryCaption extends React.Component {
 			return;
 		}
 
-		this.props.updateMedia( { siteId, item: Object.assign( {}, item, { caption } ) } );
+		this.props.updateMedia( siteId, Object.assign( {}, item, { caption } ) );
 	};
 
 	render() {
@@ -64,6 +64,7 @@ class EditorMediaModalGalleryCaption extends React.Component {
 				onChange={ this.setCaption }
 				onBlur={ this.saveCaption }
 				onMouseDown={ ( event ) => event.stopPropagation() }
+				// eslint-disable-next-line wpcalypso/jsx-classname-namespace
 				className="editor-media-modal-gallery__caption"
 			/>
 		);
