+++ client/post-editor/media-modal/index.jsx
@@ -40,7 +40,8 @@ import { getEditorPostId } from 'state/editor/selectors';
 import { resetMediaModalView } from 'state/ui/media-modal/actions';
 import { setEditorMediaModalView } from 'state/editor/actions';
 import { ModalViews } from 'state/ui/media-modal/constants';
-import { editMedia, deleteMedia } from 'state/media/actions';
+import { editMedia } from 'state/media/thunks';
+import { deleteMedia } from 'state/media/actions';
 import ImageEditor from 'blocks/image-editor';
 import VideoEditor from 'blocks/video-editor';
 import MediaModalDialog from './dialog';
@@ -309,7 +310,7 @@ export class EditorMediaModal extends Component {
 			return;
 		}
 
-		this.props.editMedia( { siteId, item: { ID: item.ID, media_url: item.guid } } );
+		this.props.editMedia( siteId, { ID: item.ID, media_url: item.guid } );
 
 		this.props.onRestoreMediaHook();
 	};
@@ -338,7 +339,7 @@ export class EditorMediaModal extends Component {
 			height && { height }
 		);
 
-		this.props.editMedia( { siteId: site.ID, item } );
+		this.props.editMedia( site.ID, item );
 
 		resetAllImageEditorState();
