+++ client/state/media/actions.js
@@ -2,7 +2,7 @@
  * External dependencies
  */
 
-import { castArray, assign } from 'lodash';
+import { castArray } from 'lodash';
 
 /**
  * Internal dependencies
@@ -29,9 +29,6 @@ import {
 } from 'state/action-types';
 
 import 'state/data-layer/wpcom/sites/media';
-import getMediaItem from 'state/selectors/get-media-item';
-import { createTransientMedia } from 'lib/media/utils';
-import { dispatchFluxUpdateMediaItem } from './actions/utils';
 
 /**
  * Returns an action object used in signalling that media item(s) for the site
@@ -204,50 +201,31 @@ export function createMediaItem( site, transientMedia ) {
 	};
 }
 
+/**
+ * Returns an action object used in signalling that media item for the site
+ * are to be edited.
+ *
+ * @param {number} siteId site identifier
+ * @param {object} item media item
+ */
 export const editMediaItem = ( siteId, item ) => ( {
 	type: MEDIA_ITEM_EDIT,
 	siteId,
 	item,
 } );
 
-export const editMedia = ( action ) => ( dispatch, getState ) => {
-	const { siteId, item } = action;
-
-	const transientMediaItem = createTransientMedia( item.media || item.media_url );
-
-	if ( ! transientMediaItem ) {
-		return;
-	}
-
-	const mediaId = item.ID;
-	const originalMediaItem = getMediaItem( getState(), siteId, mediaId );
-	const editedMediaItem = assign( {}, originalMediaItem, transientMediaItem, {
-		ID: mediaId,
-		isDirty: true,
-	} );
-
-	dispatchFluxUpdateMediaItem( siteId, editedMediaItem );
-
-	dispatch( editMediaItem( action.siteId, item ) );
-};
-
+/**
+ * Returns an action object used in signalling that media item data for the site
+ * are to be updated.
+ *
+ * @param {number} siteId site identifier
+ * @param {object} item media item
+ */
 export const updateMediaItem = ( siteId, item ) => ( {
 	type: MEDIA_ITEM_UPDATE,
 	siteId,
 	item,
 } );
-
-export const updateMedia = ( action ) => ( dispatch, getState ) => {
-	const { siteId, item } = action;
-	const mediaId = item.ID;
-
-	const originalMediaItem = getMediaItem( getState(), siteId, mediaId );
-	const updatedMediaItem = assign( {}, originalMediaItem, item );
-
-	dispatchFluxUpdateMediaItem( siteId, updatedMediaItem );
-
-	dispatch( updateMediaItem( action.siteId, updatedMediaItem ) );
-};
 
 /**
  * Returns an action object used in signalling that media item(s) for the site
