+++ client/state/media/thunks/index.js
@@ -1,3 +1,5 @@
 export { addMedia } from './add-media';
 export { uploadSiteIcon } from './upload-site-icon';
 export { addExternalMedia } from './add-external-media';
+export { editMedia } from './edit-media';
+export { updateMedia } from './update-media';
